Fix three MSVC portability bugs caused by 32-bit unsigned long on Windows
On Windows/MSVC, `unsigned long` is 32 bits (vs 64 bits on Linux/macOS).
This caused three distinct failures in the test suite:

1. std::stol/stoul throw out_of_range for values that fit in 64 bits
   but exceed 32-bit range. Replace with std::stoll/stoull in
   attribute.cpp and rowgroup.cpp.

2. String decode operators reserved 1024 * 2MB = 2GB in a single
   vector::reserve() call. Linux overcommits virtual memory so this
   succeeds; Windows does not, throwing bad_alloc. Replace with
   amortized doubling growth strategy across all 8 decode operators.

3. FSST12 decompressor cast the symbol table to `unsigned long*`,
   giving 4-byte stride on Windows instead of the required 8-byte
   stride. This caused symbol lookups at wrong offsets, producing
   correct-length but wrong-content strings. Fix by casting to
   `unsigned long long*` in fsst12.h.

Also removes debug instrumentation from the reader/materializer
that was added while investigating these issues.

+++ src/include/fls/cor/prm/fsst12/fsst12.h
@@ -145,7 +145,7 @@ fsst12_decompress(fsst12_decoder_t* decoder, /* IN: use this dictionary for comp
                   unsigned char*    output   /* OUT: memory buffer to put the decompressed string in. */
 ) {
 	unsigned char* __restrict__ len    = (unsigned char* __restrict__)decoder->len;
-	unsigned long* __restrict__ symbol = (unsigned long* __restrict__)decoder->symbol;
+	unsigned long long* __restrict__ symbol = (unsigned long long* __restrict__)decoder->symbol;
 	unsigned char* __restrict__ strOut = (unsigned char* __restrict__)output;
 	unsigned long posOut = 0, posIn = 0;
 #define FSST12_UNALIGNED_STORE(dst, src) memcpy((unsigned long long*)(dst), &(src), sizeof(unsigned long long))
